+++ cpu/src/standard_sor.cpp
@@ -10,7 +10,7 @@
 #include <iostream>                // std::cout
 
 /*--------- ctor / dtor matching the declarations --------------*/
-SolverStandardSOR::SolverStandardSOR(double* grid,
+SolverStandardSOR::SolverStandardSOR(float* grid,
                                      int     w,
                                      int     h,
                                      const std::string& n)
@@ -23,24 +23,24 @@ SolverStandardSOR::~SolverStandardSOR() = default;
 void SolverStandardSOR::solve(const SimulationParameters& prm)
 {
     const int    itMax = prm.max_iterations;
-    const double tol   = prm.tolerance;
-    const double omega = prm.omega;
+    const float tol   = prm.tolerance;
+    const float omega = prm.omega;
 
     for (int iter = 0; iter < itMax; ++iter)
     {
-        double maxErr = 0.0;
+        float maxErr = 0.0f;
 
         for (int j = 1; j < height - 1; ++j)
             for (int i = 1; i < width - 1; ++i)
             {
                 const int idx   = i + j * width;
-                const double old = U[idx];
+                const float old = U[idx];
 
-                const double sigma =
+                const float sigma =
                       ( U[idx - 1]     + U[idx + 1]
-                      + U[idx - width] + U[idx + width] ) * 0.25;
+                      + U[idx - width] + U[idx + width] ) * 0.25f;
 
-                const double diff = sigma - old;
+                const float diff = sigma - old;
                 U[idx]  += omega * diff;                 // SOR update
                 maxErr   = std::max(maxErr, std::abs(diff));
             }
